folder behaviours to behaviors
