+++ code/planning/src/behavior_agent/behaviors/intersection.py
@@ -6,7 +6,7 @@ import rospy
 import sys
 import os
 
-from behaviours import behavior_speed as bs
+from behaviors import behavior_speed as bs
 
 sys.path.append(os.path.abspath(sys.path[0] + "/.."))
 from local_planner.utils import (  # type: ignore # noqa: E402
@@ -29,6 +29,93 @@ def get_color(state):
         return "yellow"
     else:
         return ""
+
+
+class Ahead(py_trees.behaviour.Behaviour):
+    """
+    This behaviour checks whether there is an intersection in front of the
+    ego vehicle or not and triggers the rest of the decision tree handling the
+     intersection.
+    """
+
+    def __init__(self, name):
+        """
+        Minimal one-time initialisation. A good rule of thumb is to only
+        include the initialisation relevant for being able to insert this
+        behaviour in a tree for offline rendering to dot graphs.
+
+         :param name: name of the behaviour
+        """
+        super(Ahead, self).__init__(name)
+
+    def setup(self, timeout):
+        """
+        Delayed one-time initialisation that would otherwise interfere with
+        offline rendering of this behaviour in a tree to dot graph or
+        validation of the behaviour's configuration.
+
+        This initializes the blackboard to be able to access data written to it
+        by the ROS topics.
+        :param timeout: an initial timeout to see if the tree generation is
+        successful
+        :return: True, as the set up is successful.
+        """
+        self.blackboard = py_trees.blackboard.Blackboard()
+        return True
+
+    def initialise(self):
+        """
+        When is this called?
+            The first time your behaviour is ticked and anytime the status is
+            not RUNNING thereafter.
+        What to do here?
+            Any initialisation you need before putting your behaviour to work.
+        This initializes the variables needed to save information about the
+        stop line.
+        """
+        self.dist = 0
+        return True
+
+    def update(self):
+        """
+        When is this called?
+        Every time your behaviour is ticked.
+        What to do here?
+            - Triggering, checking, monitoring. Anything...but do not block!
+            - Set a feedback message
+            - return a py_trees.common.Status.[RUNNING, SUCCESS, FAILURE]
+        Gets the current distance to the next intersection.
+        :return: py_trees.common.Status.SUCCESS, if the vehicle is within range
+                    of the intersection
+                 py_trees.common.Status.FAILURE, if we are too far away from
+                 the intersection
+        """
+
+        bb = self.blackboard.get("/paf/hero/waypoint_distance")
+        if bb is None:
+            return py_trees.common.Status.FAILURE
+        else:
+            dist = bb.distance
+            isIntersection = bb.isStopLine
+        if dist < 30 and isIntersection:
+            return py_trees.common.Status.SUCCESS
+        else:
+            return py_trees.common.Status.FAILURE
+
+    def terminate(self, new_status):
+        """
+        When is this called?
+        Whenever your behaviour switches to a non-running state.
+            - SUCCESS || FAILURE : your behaviour's work cycle has finished
+            - INVALID : a higher priority branch has interrupted, or shutting
+            down
+        writes a status message to the console when the behaviour terminates
+        :param new_status: new state after this one is terminated
+        """
+        self.logger.debug(
+            "  %s [Foo::terminate().terminate()][%s->%s]"
+            % (self.name, self.status, new_status)
+        )
 
 
 class Approach(py_trees.behaviour.Behaviour):
