+++ code/planning/src/behavior_agent/behaviors/maneuvers.py
@@ -2,7 +2,7 @@ import py_trees
 import rospy
 from std_msgs.msg import String, Float32, Bool
 import numpy as np
-from behaviours import behavior_speed as bs
+from behaviors import behavior_speed as bs
 
 # from behavior_agent.msg import BehaviorSpeed
